fix(window): render windows correctly when partially off-screen (#9)
* docs(plans): add plan for partial off-screen window rendering fix

* fix(window): clip partial off-screen windows in signed coordinates

Migrate the GraphicsDevice trait's drawing primitives from `usize` to
`i32` (positions) and `u32` (sizes), and clip in signed space at the
adapter boundary. This fixes the wrap-around and pixel corruption that
appeared when a window was dragged so its bounds extended past a screen
edge: callers used to cast negative `bounds.x` through `as usize`,
producing huge wrapped values that re-wrapped during fill/line math and
landed pixels at unrelated screen offsets.

The clip helper in `src/window/adapters/clip.rs` widens to `i64` for
overflow safety, intersects rects against device bounds and the active
clip rect, and provides a Cohen–Sutherland line clipper so Bresenham can
work over already-bounded endpoints. Adapters call the helper once per
primitive; `draw_text` pre-clips per glyph and skips per-pixel checks
when a glyph is fully inside the clipped region.

CursorRenderer's public API moves to `i32` so cursor positions don't
re-introduce the unsigned trap at the manager-cursor seam. The
underlying `FrameBufferWriter` keeps its `usize` interface unchanged —
adapters guard the boundary.

Sweeps every window painter (frame, text, desktop, container, button,
label, list, menu, menu_bar, menu_bar_popup, multi_column_list, taskbar,
text_input, text_editor) plus the painting command and manager cursor
call sites to drop `as usize` casts; coordinates flow through `i32` end
to end.

* feat(window): clamp drag so the title bar stays grabbable

After the partial-off-screen rendering fix, a window can be dragged
arbitrarily far off any edge. Add a lenient drag-time clamp so at least
80 px of the title bar remains within the screen (or the whole window
if it is narrower than that), and the title bar's vertical band stays
inside [0, screen_height - 24]. Partial off-screen drag is otherwise
unrestricted.

The clamp lives in `clamp_drag_x` / `clamp_drag_y` free functions in
`src/window/types.rs` so it can be exercised by the new in-kernel
tests, alongside coverage for the signed-coordinate clipping helpers
(`clip_rect`, `pixel_visible`, `clip_line`) that landed with U1.

* docs(plans): mark partial off-screen rendering plan completed

+++ src/window/windows/desktop.rs
@@ -57,10 +57,10 @@ impl Window for DesktopWindow {
         // Fill the entire desktop with the background color
         let bounds = self.base.bounds();
         device.fill_rect(
-            bounds.x as usize,
-            bounds.y as usize,
-            bounds.width as usize,
-            bounds.height as usize,
+            bounds.x,
+            bounds.y,
+            bounds.width,
+            bounds.height,
             self.background_color,
         );
